# api/DTOs/Inputs/MedicationStorageMissingInputDto.cs
using System;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;

namespace hospitalApi.DTOs.Inputs;

public partial class MedicationStorageMissingInputDto
{
    [DefaultValue(1)]
    public int FkMedicationStorageId { get; set; }

    [DefaultValue(10)]
    public double AmountMissing { get; set; }

    [DefaultValue("2025-10-07T12:00:00")]
    public DateTime WentMissingAt { get; set; }
}