# api/swagger/v1/swagger.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "hospitalApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Medicin": {
      "get": {
        "tags": [
          "Medicin"
        ],
        "operationId": "Medicin_GetAllMedicins",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationOutputDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationOutputDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationOutputDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Medicin"
        ],
        "operationId": "Medicin_Post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/Medicin/{id}": {
      "get": {
        "tags": [
          "Medicin"
        ],
        "operationId": "Medicin_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationOutputDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationOutputDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Medicin"
        ],
        "operationId": "Medicin_Put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Medicin"
        ],
        "operationId": "Medicin_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/MissingStorage": {
      "get": {
        "tags": [
          "MissingStorage"
        ],
        "operationId": "MissingStorage_GetAllMedicationStorageMissings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageMissingOutputDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageMissingOutputDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageMissingOutputDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MissingStorage"
        ],
        "operationId": "MissingStorage_Post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageMissingInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageMissingInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageMissingInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/MissingStorage/{id}": {
      "get": {
        "tags": [
          "MissingStorage"
        ],
        "operationId": "MissingStorage_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageMissingOutputDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageMissingOutputDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageMissingOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "MissingStorage"
        ],
        "operationId": "MissingStorage_Put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageMissingInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageMissingInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageMissingInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "MissingStorage"
        ],
        "operationId": "MissingStorage_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Storage/{id}": {
      "get": {
        "tags": [
          "Storage"
        ],
        "operationId": "Storage_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageOutputDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageOutputDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Storage"
        ],
        "operationId": "Storage_Put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Storage"
        ],
        "operationId": "Storage_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "hospitalApi.DTOs.Inputs.MedicationInputDto": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "genericName": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "brand": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "form": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "strength": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "category": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "hospitalApi.DTOs.Inputs.MedicationStorageInputDto": {
        "type": "object",
        "properties": {
          "fkMedicationId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "hospitalApi.DTOs.Inputs.MedicationStorageMissingInputDto": {
        "type": "object",
        "properties": {
          "fkMedicationStorageId": {
            "type": "integer",
            "format": "int32"
          },
          "amountMissing": {
            "type": "number",
            "format": "double"
          },
          "wentMissingAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "hospitalApi.DTOs.Outputs.MedicationOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "genericName": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "brand": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "form": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "strength": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "category": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "hospitalApi.DTOs.Outputs.MedicationStorageMissingOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fkMedicationStorageId": {
            "type": "integer",
            "format": "int32"
          },
          "amountMissing": {
            "type": "number",
            "format": "double"
          },
          "wentMissingAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "hospitalApi.DTOs.Outputs.MedicationStorageOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fkMedicationId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
...
}